+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix [#163](https://github.com/squint-cljs/cherry/issues/163): Add assert to macros
 - Fix [#165](https://github.com/squint-cljs/cherry/issues/165): Fix ClojureScript protocol dispatch
 - Fix [#169](https://github.com/squint-cljs/cherry/issues/169): `extend-type` on `Object`
+- Fix [#171](https://github.com/squint-cljs/cherry/issues/171): Add `satisfies?` macro
 
 ## 0.4.32 (2025-10-21)